backend/requirements.txt
Django>=4.2
 djangorestframework
 requests
 django-cors-headers 